requirements.txt
playwright==1.40.0
requests
anthropic
Pillow
python-dotenv
